% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFussionInput.R
\name{calcFussionInput}
\alias{calcFussionInput}
\title{FussionInput}
\usage{
calcFussionInput()
}
\value{
A list of input reuslts
}
\description{
FussionInput
}
\examples{

\dontrun{
a <- calcOutput(type = "FussionInput", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
