% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFussionOutput.R
\name{calcFussionOutout}
\alias{calcFussionOutout}
\title{Calculate Democratic Republic of the Congo model outputs
Reads the Democratic Republic of the Congo source data, converts the model
results to a quitte data frame, and maps OSeMOSYS technologies, commodities,}
\usage{
calcFussionOutout()
}
\value{
A list of DRC reuslts
}
\description{
emissions, costs, capacity, generation, demand, and activity variables to
IAMC-style variable names.
}
\details{
Technology and commodity descriptions are obtained from the SETS table.
The resulting data frame retains the original model dimensions and adds
descriptive technology, fuel, and IAMC variable columns.
}
\examples{

\dontrun{
a <- calcOutput(type = "FussionOutout")
}

}
\author{
Fotis Sioutas
}
