% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDemocraticRepublicCongo.R
\name{downloadDemocraticRepublicCongo}
\alias{downloadDemocraticRepublicCongo}
\alias{readDemocraticRepublicCongo}
\title{Read-in electricity production and energy balance data for DemocraticRepublicCongo}
\usage{
downloadDemocraticRepublicCongo()

readDemocraticRepublicCongo()
}
\value{
The read-in data as a MAgPIE object.
}
\description{
Reads RE-INTEGRATE datasets containing electricity production by sector
and national energy balance statistics for DemocraticRepublicCongo.
}
\examples{
\dontrun{
DRC <- readSource("DemocraticRepublicCongo")
}

}
\author{
Fotis Sioutas
}
