% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readElecProd.R
\name{downloadElecProd}
\alias{downloadElecProd}
\alias{readElecProd}
\alias{convertElecProd}
\title{Read-in data from RE-INTEGRATE about Electricity Production}
\usage{
downloadElecProd()

readElecProd()

convertElecProd(x)
}
\arguments{
\item{x}{MAgPIE object returned by readElecProd}
}
\value{
The read-in data into a magpie object.
}
\description{
Read-in data from RE-INTEGRATE about Electricity Production
}
\examples{
\dontrun{
a <- readSource("ElecProd", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
