% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readETHIOPIA.R
\name{readETHIOPIA}
\alias{readETHIOPIA}
\title{Reads RE-INTEGRATE datasets for ETHIOPIA}
\usage{
readETHIOPIA()
}
\value{
A MAgPIE object containing the Ethiopian model results.
}
\description{
Reads model results for the Baseline and Ambitious_85percentBy2040 scenarios
}
\author{
Fotis Sioutas #' #' @examples

\dontrun{
ETHIOPIA <- readSource("ETHIOPIA")
}
}
