% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAURITIUS.R
\name{readMAURITIUS}
\alias{readMAURITIUS}
\title{Reads RE-INTEGRATE datasets for MAURITIUS}
\usage{
readMAURITIUS()
}
\value{
A list containing a MAgPIE-compatible quitte object and metadata.
}
\description{
Reads model results for the Test293 and Test292 scenarios.
All CSV files are combined into one data frame. Columns that are missing
from individual CSV files are automatically filled with NA. The original
name of the last column in each CSV file is stored in the variable column.
}
\examples{
\dontrun{
MUS <- readSource("MAURITIUS")
}

}
\author{
Fotis Sioutas
}
