% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSouthAfrica.R
\name{downloadSouthAfrica}
\alias{downloadSouthAfrica}
\alias{downSouthAfrica}
\alias{readSouthAfrica}
\title{Read-in electricity production and energy balance data for South Africa}
\usage{
downloadSouthAfrica()

downSouthAfrica()

readSouthAfrica()
}
\value{
The read-in data.
}
\description{
Reads RE-INTEGRATE datasets containing electricity production by sector
and national energy balance statistics for South Africa
}
\examples{
\dontrun{
SouthAfrica <- readSource("SouthAfrica")
}

}
\author{
Fotis Sioutas
}
