% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readElecProd.R
\name{downloadElecProd}
\alias{downloadElecProd}
\alias{readElecProd}
\title{Read-in data from RE-INTEGRATE about Electricity Production}
\usage{
downloadElecProd()

readElecProd()
}
\value{
The read-in data into a magpie object.
}
\description{
Read-in data from RE-INTEGRATE about Electricity Production
}
\examples{
\dontrun{
a <- readSource("ElecProd")
}

}
\author{
Fotis Sioutas
}
