% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSenegal.R
\name{downloadSenegal}
\alias{downloadSenegal}
\alias{readSenegal}
\title{Read-in electricity production and energy balance data for Senegal.}
\usage{
downloadSenegal()

readSenegal(subtype = "EleProd")
}
\arguments{
\item{subtype}{Character string specifying the dataset to import.
Supported values are "EleProd" and "EnergyBalances".}
}
\value{
The read-in data as a MAgPIE object.
}
\description{
Reads RE-INTEGRATE datasets containing electricity production by sector
and national energy balance statistics for Senegal. Depending on the
selected subtype, the function imports and formats either electricity
production data ("EleProd") or detailed energy balance data
("EnergyBalances"), returning the results as a MAgPIE object.
}
\examples{
\dontrun{
elec <- readSource("Senegal", subtype = "EleProd")
balances <- readSource("Senegal", subtype = "EnergyBalances")
}

}
\author{
Fotis Sioutas
}
