% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSenegal.R
\name{downloadSenegal}
\alias{downloadSenegal}
\alias{readSenegal}
\title{Read-in electricity production and energy balance data for Senegal.}
\usage{
downloadSenegal()

readSenegal()
}
\value{
The read-in data.
}
\description{
Reads RE-INTEGRATE datasets containing electricity production by sector
and national energy balance statistics for Senegal. The function imports
and formats for electricity production data or detailed energy balance data.
}
\examples{
\dontrun{
Senegal <- readSource("Senegal")
}

}
\author{
Fotis Sioutas
}
